requirements.txt
langgraph~=0.5.3
typing_extensions~=4.14.1
langchain~=0.3.26
python-dotenv~=1.1.1
pydantic~=2.11.7